Float To BigDecimal

+++ src/main/java/com/joannava/kafka/katas/model/CountAndSum.java
@@ -1,5 +1,7 @@
 package com.joannava.kafka.katas.model;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
@@ -9,5 +11,5 @@ import lombok.extern.jackson.Jacksonized;
 @Data
 public class CountAndSum {
     private final long count;
-    private final float sum;
+    private final BigDecimal sum;
 }
